+++ client/assets/components/document/document_jira/contentTypes/jiraProject.js
@@ -24,32 +24,18 @@
   }
 
 
-  function Controller(SignalsService, PaginateService) {
+  function Controller(DocumentService) {
     'ngInject';
     var vm = this;
-
-    activate();
-
-    function activate() {
-      vm.postSignal = postSignal;
-      vm.doc = processDocument(vm.doc);
-    }
-
-    function processDocument(doc) {
-      doc.__signals_doc_id__ = SignalsService.getSignalsDocumentId(doc);
-      doc.page = PaginateService.getNormalizedCurrentPage();
-      return doc;
-    }
+    vm.postSignal = postSignal;
+    vm.getTemplateDisplayFieldName = getTemplateDisplayFieldName;
 
     function postSignal(options){
-      var paramsObj = {
-        params: {
-          position: vm.doc.position,
-          page: vm.doc.page
-        }
-      };
-      _.defaultsDeep(paramsObj, options);
-      SignalsService.postClickSignal(vm.doc.__signals_doc_id__, paramsObj);
+      DocumentService.postSignal(vm.doc._signals, options);
+    }
+
+    function getTemplateDisplayFieldName(field){
+      return DocumentService.getTemplateDisplayFieldName(vm.doc, field);
     }
   }
 })();